Add interactive functionality to RecidivizSlider component
