+++ src/components/App/App.css
@@ -5,14 +5,6 @@ main {
 	min-height: 100vh;
 }
 
-svg .first-chart {
-   overflow:none;
-   padding:10px;
-   float:left;
-   width:600px;
-   height:600px;
-}
-
 .bg-alert-blue {
  background-color:#eeeeee;
 }
@@ -21,47 +13,10 @@ svg .first-chart {
   background-color:#d6d4d3;
 }
 
-.iconPlain  {
-	fill:#c5c2c2;
-
-}
-
-.iconSelected  {
-	fill:#c82222;
-}
-
-.nunito {
-	font-family: 'Nunito', monospace;
-}
-
-rect    {
-  fill:#ffffff;
-}
-
 .text-our-gray {
 	color: #766e6e;
 }
 
 .text-our-highlight {
 	color: #c82222;
-}
-
-.tooltip {
-   position: absolute;
-   z-index: 1020;
-   padding: 5px;
-   width: 200px;
-   font-size: 36px;
-   visibility: visible;
-   margin-top: -2px;
-   bottom:120%;
-   margin-left: -2em;
-}
-
-#sliderDiv  {
-   margin:auto;
-   margin-top:30px;
-   height:15px;
-   width:50%;
-   text-align: center;
 }
