+++ src/components/App/App.test.js
@@ -2,6 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 
+jest.mock('../../lib/static/recidivismSlider.js');
+
 it('renders the App component', () => {
   const div = document.createElement('div');
   ReactDOM.render(<App />, div);
