+++ src/components/Recidiviz/RecidivizSlider/RecidivizSlider.css
@@ -1,0 +1,46 @@
+#sliderDiv {
+   height: 15px;
+   width: 50%;
+   text-align: center;
+}
+
+#sliderDiv .ui-state-active {
+  background-image: none;
+  background: #ffffff 50% 50%;
+  border-color: #aaaaaa;
+}
+
+#txtValue {
+  font-family: 'Nunito', monospace;
+  margin-bottom: 0;
+  height: 77px;
+}
+
+.first-chart svg {
+  overflow: none;
+  padding: 10px;
+  float: left;
+  width: 600px;
+  height: 600px;
+}
+
+.iconPlain {
+  fill: #c5c2c2;
+}
+
+.iconSelected {
+  fill: #c82222;
+}
+
+rect {
+  fill: #ffffff;
+}
+
+.ui-slider-handle {
+  outline: none;
+}
+
+h1 {
+  margin-top: 10px;
+  margin-bottom: 10px;
+}
