+++ src/config/jestSetup.js
@@ -1,0 +1,10 @@
+import $ from 'jquery';
+
+global.$ = $;
+global.jQuery = $;
+
+require('jquery-ui/ui/version');
+require('jquery-ui/ui/plugin');
+require('jquery-ui/ui/widget');
+require('jquery-ui/ui/widgets/mouse');
+require('jquery-ui/ui/widgets/slider');
